use array of button pointers to simplify code
Rather than checking each button individually, we can loop through the
buttons and check each of them. This required the addition of an
additional state variable to the button class, for what screen it led
to. (This may make the button class slightly less flexible, but we can
work with that in the future if we need buttons somewhere else. Maybe
move this functionality into a subclass.)