finish quitting support
Date code works properly and the interface code works properly. The one
remaining possible issue is that if you stop the quit button when it's
right in the middle of the screen, it's impossible to see the date to
answer it properly. Might want to try to fix that somehow.